run initializer on all replicas
This modifies the model of our bootstrapping so that the initializer runs on each replica and
configures only the replica which it runs on.

I think this has a number of benefits:
* Supports scaling replicas. If you increase the number of replicas, a new replica will join
itself to the current leader automatically.
* Removes the need to provide the number of replicas to the initializer.
* If a replica in the statefulset becomes unrecoverable or reverts to an uninitialized
state - perhaps due to losing its pvc - it will be returned to the cluster on startup.
* When initialization fails, for whatever reason, initialization can be reattempted by
restarting the affected pod.