export metrics with the statsd-exporter (#30)
The vault prometheus metrics are less than ideal for a couple of reasons:
* They're event based, and metrics are only exported when a particular event
happens and persisted for as long as `prometheus_retention_time`. Additionally,
counters seem to be reset every time a metric reappears, meaning you're
essentially getting the rate as a gauge. We could increase the retention time to
mitigate this a bit (it's currently set to `90s`).
* The conversion from statsd format to prometheus format is done naively,
resulting in information being placed into metrics names that should probably go
into labels. For instance `vault_rollback_attempt_aws_` and
`vault_rollback_attempt_cubbyhole_` which could be
`vault_rollback_attempt{path="aws"}` and `vault_rollback_attempt{path="cubbyhole"}`.

The statsd-exporter performs the exact same conversion from statsd to prometheus
but metrics received from vault are persisted for the life of the exporter,
meaning that counters are incremented as expected and series will generally
persist for the lifetime of the pod. It also allows you to map parts of the
statsd metric names to labels, resulting in more useful metrics.

I've mapped some metrics that stood out to me, but I'd like to follow up by
reviewing them all more thoroughly.