vault-toolkit: switch to bash
We need to use the pipefail option in the initialiser script: in the
`curl | jq` checks, if curl fails because the endpoint is not available
yet, jq does not produce an error with nil input. Therefore, when the
container starts, it thinks that vault is initialised and goes to sleep.

Also, fix various small issues detected by shellcheck.

This enables nounset, errexit and pipefail universally across all
scripts which might require more changes but is better to have them
enabled, to avoid similar issues in the future.

Signed-off-by: Dimitrios Karagiannis <dhkarag@gmail.com>